fase3-pi/requirements.txt
# Fase 3 Pi Client Dependencies
# Getest op Pi 5 met Python 3.13, Miniforge conda

# Wake word
openwakeword==0.4.0  # v0.6.0 faalt door tflite-runtime

# Audio
pyaudio
numpy

# VAD (Silero via ONNX, geen PyTorch!)
onnxruntime  # Automatisch geinstalleerd via openwakeword

# WebSocket client
websockets>=12.0

# Camera Module 3 (IMX708)
picamera2
pillow  # Voor image processing

# OLED Display
luma.oled
